requirements.txt
numpy
pillow
opencv-python
matplotlib
scikit-image
scipy
tqdm
seaborn
pytest
requests